Day 17 - OMG I can't believe i managed it
Whaaat. That was crazy. I feel very proud of myself for managing to work it out. Well done Brian!

First I did a few things that were fun but didn't really help.

I wrote out my program in code and realised it was essentially a while loop that decreased a by a factor of 8 in each iteration till it reached zero, adding one output each time.

So for our 16 character output it had to be a number between 8^15 and 8^16. Unfortunately that still left a big search space.

Simplifying the program once I'd written it down got me to this simpler form having eliminated b and c.

    while a != 0:
        out.append(((a % 8) ^ (a // 2 ** ((a % 8) ^ 1)) % 8) ^ 4)
        a = a // 8

Then I think I took the same approach as you Brian. I took the last output (16th) and inferred the previous one based on
a // 8 had to be zero, so a could be from 0 to 7
Test those values individually using the expression from above
((a % 8) ^ (a // 2 ** ((a % 8) ^ 1)) % 8) ^ 4 = n
Where n was the value of the last output. This gave me possible value/s of a 1 output before the program completed.

Then I looked at the next digit left, the 15th.
a // 8 has to one of the values I calculated above z, so it had to be from  z * 8 to z * 8 + 7
Again I tested those 8 values individually to get all the possible values of a 2 outputs before the program completed

I then had to make this a dfs since there were digits that had multiple solutions but some of them resulted in problem that had no solution further down the line.

I also had to take care to make sure my dfs searched for solutions in increasing order since the problem statement wanted the smallest a which satisfied the condition.

And when it worked..... I couldn't believe it.

I don't intend to refactor anything here. I am quite proud of the complete chaos that was my brain wrestling to find the right number and not thinking about coding at all.